[#133] refactor: Magin number 상수로 대체
ConstantsCoveringMagicNumber
- 매직 넘버의 사용을 지양하기 위해 해당 데이터들을 모두 static 상수화
- ADMIN_PAGE_SIZE : ADMIN에서 사용하는 페이징 기본 크기
- PAGE_GROUP_SIZE : 페이지 그룹 사이즈
- MINIMUM_PAGE_NUMBER : 페이지 요청 파라미터 최소값

PagingProperties
- 10 -> PAGE_GROUP_SIZE (10) 변경

BookAdminService
- 10 -> ADMIN_PAGE_SIZE (10) 변경
- 1 -> MINIMUM_PAGE_NUMBER (1) 변경

BooksWithPagingResponse,
BookAdminServiceTest
- PagingProperties 패키지 util -> common 변경
