+++ src/main/java/kr/codesquad/library/admin/common/ConstantsCoveringMagicNumber.java
@@ -1,0 +1,8 @@
+package kr.codesquad.library.admin.common;
+
+public final class ConstantsCoveringMagicNumber {
+
+    public static final int ADMIN_PAGE_SIZE = 10;
+    public static final int PAGE_GROUP_SIZE = 10;
+    public static final int MINIMUM_PAGE_NUMBER = 1;
+}
