+++ src/main/java/kr/codesquad/library/admin/service/BookAdminService.java
@@ -3,7 +3,7 @@ package kr.codesquad.library.admin.service;
 import kr.codesquad.library.admin.domain.book.BookAdminRepository;
 import kr.codesquad.library.admin.domain.book.BookSummary;
 import kr.codesquad.library.admin.domain.book.BooksWithPagingResponse;
-import kr.codesquad.library.admin.util.PagingProperties;
+import kr.codesquad.library.admin.common.PagingProperties;
 import kr.codesquad.library.domain.book.Book;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -14,6 +14,9 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static kr.codesquad.library.admin.common.ConstantsCoveringMagicNumber.ADMIN_PAGE_SIZE;
+import static kr.codesquad.library.admin.common.ConstantsCoveringMagicNumber.MINIMUM_PAGE_NUMBER;
+
 @RequiredArgsConstructor
 @Service
 @Transactional(readOnly = true)
@@ -22,7 +25,7 @@ public class BookAdminService {
     private final BookAdminRepository bookAdminRepository;
 
     public BooksWithPagingResponse findAllBooks(int page) {
-        Page<Book> books = bookAdminRepository.findAllWithCategory(PageRequest.of(validatePageNumber(page), 10));
+        Page<Book> books = bookAdminRepository.findAllWithCategory(PageRequest.of(validatePageNumber(page), ADMIN_PAGE_SIZE));
         List<Book> bookEntities = books.getContent();
         List<BookSummary> bookSummaries = bookEntities.stream()
                                                       .map(BookSummary::from)
@@ -32,7 +35,7 @@ public class BookAdminService {
     }
 
     private int validatePageNumber(int page) {
-        if (page < 1) { page = 1; }
+        if (page < MINIMUM_PAGE_NUMBER) { page = MINIMUM_PAGE_NUMBER; }
         return page - 1;
     }
 }
