+++ src/test/java/kr/codesquad/library/admin/service/BookAdminServiceTest.java
@@ -3,8 +3,7 @@ package kr.codesquad.library.admin.service;
 import kr.codesquad.library.admin.domain.book.BookAdminRepository;
 import kr.codesquad.library.admin.domain.book.BookSummary;
 import kr.codesquad.library.admin.domain.book.BooksWithPagingResponse;
-import kr.codesquad.library.admin.util.PagingProperties;
-import org.junit.jupiter.api.Test;
+import kr.codesquad.library.admin.common.PagingProperties;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.springframework.beans.factory.annotation.Autowired;
